Update README.md
Mathematics is often used to represent and analyse the relationship between two properties. For example, consider a car travelling along a road from some starting point. There is a relationship between the (time) that has elapsed since the start of the journey and the (distance) the car has travelled. This relationship ca be investigated "mathematically.